requirements.txt
numpy
pillow
spectral
torch
torchvision
matplotlib
tqdm
